Split up value handling into its own file
