+++ opencog/persist/sql/multi-driver/CMakeLists.txt
@@ -4,6 +4,7 @@
 ADD_LIBRARY (persist-sql
 	llapi
 	SQLAtomStorage
+	SQLValues
 	SQLPersistSCM
 )
